First prototype of Tableau site
A new directory NEWSITE was used to prototype a Tableau site.

Under this directory, there is a newly generated site.  Markdown
content, images, CSS and JS assets were copied from the Jekyll
site to this new site.

The new site runs now, as a proof-of-concept prototype.  No
show-stoppers are seen so far.